+++ examples/basic.py
@@ -4,8 +4,10 @@ from pydantic import BaseModel
 
 from async_fastapi_jwt_auth import AuthJWT
 from async_fastapi_jwt_auth.exceptions import AuthJWTException
+from async_fastapi_jwt_auth.auth_jwt import AuthJWTBearer
 
 app = FastAPI()
+auth_dep = AuthJWTBearer()
 
 
 class User(BaseModel):
@@ -36,22 +38,22 @@ def authjwt_exception_handler(request: Request, exc: AuthJWTException):
 # function is used to actually generate the token to use authorization
 # later in endpoint protected
 @app.post("/login")
-async def login(user: User, Authorize: AuthJWT = Depends()):
+async def login(user: User, authorize: AuthJWT = Depends(auth_dep)):
     if user.username != "test" or user.password != "test":
         raise HTTPException(status_code=401, detail="Bad username or password")
 
     # subject identifier for whom this token is for example id or username from database
-    access_token = await Authorize.create_access_token(subject=user.username)
+    access_token = await authorize.create_access_token(subject=user.username)
     return {"access_token": access_token}
 
 
 # protect endpoint with function jwt_required(), which requires
 # a valid access token in the request headers to access.
 @app.get("/user")
-async def user(Authorize: AuthJWT = Depends()):
-    await Authorize.jwt_required()
+async def user(authorize: AuthJWT = Depends(auth_dep)):
+    await authorize.jwt_required()
 
-    current_user = await Authorize.get_jwt_subject()
+    current_user = await authorize.get_jwt_subject()
     return {"user": current_user}