+++ examples/freshness.py
@@ -4,8 +4,10 @@ from pydantic import BaseModel
 
 from async_fastapi_jwt_auth import AuthJWT
 from async_fastapi_jwt_auth.exceptions import AuthJWTException
+from async_fastapi_jwt_auth.auth_jwt import AuthJWTBearer
 
 app = FastAPI()
+auth_dep = AuthJWTBearer()
 
 
 class User(BaseModel):
@@ -29,7 +31,7 @@ def authjwt_exception_handler(request: Request, exc: AuthJWTException):
 
 # Standard login endpoint. Will return a fresh access token and a refresh token
 @app.post("/login")
-async def login(user: User, Authorize: AuthJWT = Depends()):
+async def login(user: User, authorize: AuthJWT = Depends(auth_dep)):
     if user.username != "test" or user.password != "test":
         raise HTTPException(status_code=401, detail="Bad username or password")
 
@@ -39,31 +41,31 @@ async def login(user: User, Authorize: AuthJWT = Depends()):
     As we just verified their username and password, we are
     going to mark the token as fresh here.
     """
-    access_token = await Authorize.create_access_token(
+    access_token = await authorize.create_access_token(
         subject=user.username, fresh=True
     )
-    refresh_token = await Authorize.create_refresh_token(subject=user.username)
+    refresh_token = await authorize.create_refresh_token(subject=user.username)
     return {"access_token": access_token, "refresh_token": refresh_token}
 
 
 @app.post("/refresh")
-async def refresh(Authorize: AuthJWT = Depends()):
+async def refresh(authorize: AuthJWT = Depends(auth_dep)):
     """
     Refresh token endpoint. This will generate a new access token from
     the refresh token, but will mark that access token as non-fresh,
     as we do not actually verify a password in this endpoint.
     """
-    await Authorize.jwt_refresh_token_required()
+    await authorize.jwt_refresh_token_required()
 
-    current_user = await Authorize.get_jwt_subject()
-    new_access_token = await Authorize.create_access_token(
+    current_user = await authorize.get_jwt_subject()
+    new_access_token = await authorize.create_access_token(
         subject=current_user, fresh=False
     )
     return {"access_token": new_access_token}
 
 
 @app.post("/fresh-login")
-async def fresh_login(user: User, Authorize: AuthJWT = Depends()):
+async def fresh_login(user: User, authorize: AuthJWT = Depends(auth_dep)):
     """
     Fresh login endpoint. This is designed to be used if we need to
     make a fresh token for a user (by verifying they have the
@@ -74,7 +76,7 @@ async def fresh_login(user: User, Authorize: AuthJWT = Depends()):
     if user.username != "test" or user.password != "test":
         raise HTTPException(status_code=401, detail="Bad username or password")
 
-    new_access_token = await Authorize.create_access_token(
+    new_access_token = await authorize.create_access_token(
         subject=user.username, fresh=True
     )
     return {"access_token": new_access_token}
@@ -82,17 +84,17 @@ async def fresh_login(user: User, Authorize: AuthJWT = Depends()):
 
 # Any valid JWT access token can access this endpoint
 @app.get("/protected")
-async def protected(Authorize: AuthJWT = Depends()):
-    await Authorize.jwt_required()
+async def protected(authorize: AuthJWT = Depends(auth_dep)):
+    await authorize.jwt_required()
 
-    current_user = await Authorize.get_jwt_subject()
+    current_user = await authorize.get_jwt_subject()
     return {"user": current_user}
 
 
 # Only fresh JWT access token can access this endpoint
 @app.get("/protected-fresh")
-async def protected_fresh(Authorize: AuthJWT = Depends()):
-    await Authorize.fresh_jwt_required()
+async def protected_fresh(authorize: AuthJWT = Depends(auth_dep)):
+    await authorize.fresh_jwt_required()
 
-    current_user = await Authorize.get_jwt_subject()
+    current_user = await authorize.get_jwt_subject()
     return {"user": current_user}