+++ examples/generate_doc.py
@@ -5,8 +5,10 @@ from pydantic import BaseModel
 
 from async_fastapi_jwt_auth import AuthJWT
 from async_fastapi_jwt_auth.exceptions import AuthJWTException
+from async_fastapi_jwt_auth.auth_jwt import AuthJWTBearer
 
 app = FastAPI()
+auth_dep = AuthJWTBearer()
 
 
 class User(BaseModel):
@@ -29,19 +31,19 @@ def authjwt_exception_handler(request: Request, exc: AuthJWTException):
 
 
 @app.post("/login")
-async def login(user: User, Authorize: AuthJWT = Depends()):
+async def login(user: User, authorize: AuthJWT = Depends(auth_dep)):
     if user.username != "test" or user.password != "test":
         raise HTTPException(status_code=401, detail="Bad username or password")
 
-    access_token = await Authorize.create_access_token(subject=user.username)
+    access_token = await authorize.create_access_token(subject=user.username)
     return {"access_token": access_token}
 
 
 @app.get("/protected", operation_id="authorize")
-async def protected(Authorize: AuthJWT = Depends()):
-    await Authorize.jwt_required()
+async def protected(authorize: AuthJWT = Depends(auth_dep)):
+    await authorize.jwt_required()
 
-    current_user = await Authorize.get_jwt_subject()
+    current_user = await authorize.get_jwt_subject()
     return {"user": current_user}
 
 
@@ -65,7 +67,7 @@ def custom_openapi():
     }
 
     # Get routes from index 4 because before that fastapi define router for /openapi.json, /redoc, /docs, etc
-    # Get all router where operation_id is authorize
+    # Get all router where operation_id is "authorize"
     router_authorize = [
         route for route in app.routes[4:] if route.operation_id == "authorize"
     ]